feat: Restore landing page and add inquiry functionality
- Moves the product browsing and inquiry functionality from the `Products` page to the landing page (`Index`).
- Removes the now-redundant `Products` page and its route.
- Updates the signup flow to redirect to the new landing page.
- Disables email confirmation in the Supabase config for easier testing.
