+++ supabase/config.toml
@@ -1,1 +1,4 @@
 project_id = "cmlalcbrzrccfmpdzxml"
+
+[auth.email]
+enable_confirmations = false
